+++ apps/demo/src/app/layout.tsx
@@ -1,30 +1,11 @@
-import { contentPath, initContent } from '@/content'
+import { cms } from '@/cms'
 import { DeepDishProvider } from '@deepdish/cms'
-import { createJsonResolver } from '@deepdish/resolvers/json'
-import { configure } from '@deepdish/ui/config'
-import { typographySchema } from '@deepdish/ui/schemas'
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
+
 import './globals.css'
 
-await initContent()
-
-await configure({
-  contracts: {
-    typography: {
-      resolver: createJsonResolver(contentPath, typographySchema, {
-        maxBatchSize: 10,
-      }),
-    },
-  },
-  logging: {
-    enabled: process.env.NODE_ENV === 'development',
-  },
-  settings: {
-    baseUrl: process.env.BASE_URL,
-    draft: process.env.DEEPDISH_MODE === 'draft',
-  },
-})
+export const components = await cms()
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',