+++ packages/ui/src/config/config.ts
@@ -1,14 +1,10 @@
 import type { Result } from '@byteslice/result'
+import type { Schema } from '@deepdish/core/schema'
 import { getLogger } from '@logtape/logtape'
-import type { ValueType } from '../schemas'
-import type { Contract } from './contract'
+import type { Contract, Contracts } from './contract'
 import { configureLogging } from './logging'
 
 const logger = getLogger(['deepdish', 'config'])
-
-type Contracts = Readonly<{
-  [T in ValueType]?: Contract<T>
-}>
 
 type Logging = Readonly<{
   enabled: boolean
@@ -52,23 +48,15 @@ export async function configure(input: Config): Promise<Result<void>> {
   return { data: undefined }
 }
 
-export function getConfig(): Result<Config> {
+export function getContract(name: string): Result<Contract<Schema>> {
   if (!config) {
     return notConfigured()
   }
 
-  return { data: config }
-}
-
-export function getContract<T extends ValueType>(type: T): Result<Contract<T>> {
-  if (!config) {
-    return notConfigured()
-  }
-
-  const contract = config.contracts[type]
+  const contract = config.contracts[name]
   if (!contract) {
-    logger.error('Missing {type} contract.', { type })
-    return { failure: new Error(`Missing '${type}' contract.`) }
+    logger.error('Missing {name} contract.', { name })
+    return { failure: new Error(`Missing '${name}' contract.`) }
   }
 
   return { data: contract }