+++ packages/ui/src/elements/link.tsx
@@ -20,7 +20,6 @@ export function Link(props: ElementProps<'a', string>) {
           </a>
         )
       }}
-      type="link"
     />
   )
 }