+++ packages/workbench/lib/index.tsx
@@ -1,13 +1,18 @@
 'use client'
 
 import '../src/index.css'
+import { ShadowProvider } from '@/lib/context'
 import {
+  Workbench as InternalWorkbench,
   type WorkbenchProps as InternalWorkbenchProps,
-  Shadow,
 } from '@/workbench'
 
 type WorkbenchProps = Omit<InternalWorkbenchProps, 'ref'>
 
 export function Workbench(props: WorkbenchProps) {
-  return <Shadow {...props} />
+  return (
+    <ShadowProvider>
+      <InternalWorkbench {...props} />
+    </ShadowProvider>
+  )
 }