+++ packages/workbench/src/workbench.tsx
@@ -1,7 +1,5 @@
 import { WorkbenchProvider } from '@/lib/context'
 import { RouterProvider } from '@tanstack/react-router'
-import { type RefObject, useRef } from 'react'
-import { Tailwind } from 'react-shadow-scope'
 import { router } from './router'
 
 declare module '@tanstack/react-router' {
@@ -11,27 +9,8 @@ declare module '@tanstack/react-router' {
 }
 
 export type WorkbenchProps = {
-  ref?: RefObject<HTMLElement>
   title?: React.ReactNode
   authDisabled?: boolean
-}
-
-export function Shadow(props: WorkbenchProps) {
-  const ref = useRef<HTMLDivElement>(null)
-
-  return (
-    <Tailwind href="/__deepdish/css">
-      <div ref={ref} className="deepdish-shadow">
-        <WorkbenchProvider
-          ref={ref}
-          title={props.title}
-          authDisabled={props.authDisabled}
-        >
-          <RouterProvider router={router} />
-        </WorkbenchProvider>
-      </div>
-    </Tailwind>
-  )
 }
 
 export function Workbench(props: WorkbenchProps) {